+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+* Initial release of debounce_tap package
+* Added Debounce.inkwell widget with debounce functionality
+* Added Debounce.gestureDetector widget with debounce functionality  
+* Added Debounce.iconButton widget with debounce functionality
+* Added Debounce.textButton widget with debounce functionality
+* Added onTapBlocked callback support for InkWell
+* Added customizable debounce duration
+* Added comprehensive example app